Prepare to rename the layer decorator settings file
