+++ layer/decorators/settings.py
@@ -1,0 +1,1 @@
+from layer.settings import LayerSettings  # noqa pylint: disable=unused-import
